requirements.txt
openai==2.2.0
pyside6==6.1.3
pillow==10.4.0
pygame==2.6.1
peewee==3.18.3
setuptools==75.1.0
numpy==1.24.4
scipy==1.10.1
keyring==25.5.0
cryptography==46.0.3